requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.2
bidict==0.23.1
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
openpyxl==3.1.5
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
PySocks==1.7.1
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.5
rsa==4.9.1
selenium==4.38.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
